small improvements
