--- libs/core/src/lib/utils/css-class-builder.interface.ts
@@ -1,7 +1,0 @@
-import { ElementRef } from '@angular/core';
-
-export interface CssClassBuilder {
-    class: string // user's custom classes
-    elementRef(): ElementRef;
-    buildComponentCssClass(): string;
-}
