--- libs/core/src/lib/utils/css-style-builder.interface.ts
@@ -1,10 +1,0 @@
-import { ElementRef } from '@angular/core';
-
-export interface Hash {
-    [key: string]: any;
-}
-
-export interface CssStyleBuilder {
-    elementRef(): ElementRef;
-    buildComponentCssStyle(): Hash;
-}
