+++ libs/core/src/lib/utils/decorators/apply-css-style.decorator.ts
@@ -1,15 +1,17 @@
-import { Hash } from '../css-style-builder.interface';
+import { Hash, ELEMENT_REF_EXCEPTION } from '../public_api';
 
 export function applyCssStyle(target: any, propertyKey: string, descriptor: PropertyDescriptor) {
     const originalMethod = descriptor.value
     console.log(descriptor);
     descriptor.value = function (): Hash {
+        if (!this.elementRef) { throw ELEMENT_REF_EXCEPTION; }
+
         const _styles: Hash = originalMethod.apply(this);
-        if (this.elementRef) {
-            Object.keys(_styles).forEach(key => {
-                (this.elementRef().nativeElement as HTMLElement).style[key] = _styles[key];
-            });
-        };
+
+        Object.keys(_styles).forEach(key => {
+            (this.elementRef().nativeElement as HTMLElement).style[key] = _styles[key];
+        });
+
         return _styles;
     }
 }
