+++ libs/core/src/lib/utils/interfaces/css-class-builder.interface.ts
@@ -1,0 +1,6 @@
+import { HasElementRef } from '../public_api';
+
+export interface CssClassBuilder extends HasElementRef {
+    class: string
+    buildComponentCssClass(): string;
+}
