+++ libs/core/src/lib/utils/interfaces/css-style-builder.interface.ts
@@ -1,0 +1,9 @@
+import { HasElementRef } from '../public_api';
+
+export interface Hash {
+    [key: string]: any;
+}
+
+export interface CssStyleBuilder extends HasElementRef {
+    buildComponentCssStyle(): Hash;
+}
