+++ libs/core/src/lib/utils/interfaces/has-element-ref.interface.ts
@@ -1,0 +1,7 @@
+import { ElementRef } from '@angular/core';
+
+export interface HasElementRef {
+    elementRef(): ElementRef;
+}
+
+export const ELEMENT_REF_EXCEPTION = 'HasElementRef interface has to be implemented';
