+++ libs/core/src/lib/utils/public_api.ts
@@ -4,7 +4,8 @@ export * from './drag-and-drop/drag-and-drop.module';
 export * from './drag-and-drop/dnd-list/dnd-list.directive';
 export * from './drag-and-drop/dnd-container/dnd-container.directive';
 
-export * from './css-class-builder.interface';
-export * from './css-style-builder.interface';
+export * from './interfaces/css-class-builder.interface';
+export * from './interfaces/css-style-builder.interface';
+export * from './interfaces/has-element-ref.interface';
 export * from './decorators/apply-css-class.decorator';
 export * from './decorators/apply-css-style.decorator';
